Generalise `ModelFindVariables` and `ModelLookup` to `ModelVarContext`.
Occurrences of `ModelFindVariables` and `ModelLookup` in the `InLockstep` class
are replaced by the newly exposed `ModelVarContext`. A `ModelFindVariables` can
be recovered from a `ModelVarContext` using the new `findVars` functions. A
`ModelLookup` can be recovered from a `ModelVarContext` using the new
`lookupVars` function. Since these functions can be recovered from
`ModelVarContext`, existing tests are guaranteed to be adaptable to the new
`InLockstep` API.

Motivation: previously in the `InLockstep` class, member functions would be
passed a `ModelFindVariables` or a `ModelLookup`, but never both. In practice
this turned out to be too restrictive, because one might want access to both in
the same function, see https://github.com/IntersectMBO/lsm-tree/pull/431. For
example, `arbitraryWithVars` was previously only passed a `ModelFindVariables`,
but without a `ModelLookup` one can not filter these variables based on the
(modelled) outcome of the corresponding actions. The use case for #431 in
particular was to filter variables that reference stateful handles (e.g., file
handles) for handles that are open. Because of this filtering, one can skew the
action distribution to generate more actions on open handles. It is often more
interesting to test actions on open handles than it is to test actions on closed
handles, which will presumably always return the same error.